// src/types/types.ts
export type Image = {
  id: number;
  url: string;
};

export type Interest = {
  id: number;
  name: string;
  subInterests: Interest[];
};

export type UserStatus =
  | "AVAILABLE_TO_YAKKA"
  | "AVAILABLE_TO_CHAT"
  | "UNAVAILABLE";
// ...
export type GroupType = {
  id: number;
  name: string;
  description: string;
  date: Date;
  endTime: Date;
  coverImage?: string | null;
  image?: Image;
  locationName:string[],
  categories:string[],
  profileImage?:string| null,
  reviews: {
    average: number;
    total: number;
  };
  groupGender:string;
  isOrganiser: boolean;
  hashtags: string[];
  isMember: boolean;
  isInvited: boolean;
  isPrivate: boolean;
  paymentAmount: number;
  paymentUrl: string;
};